Move Header into seperate component
