+++ frontend/src/components/Main/mainStyle.js
@@ -17,15 +17,3 @@ export const MainContainer = styled('main')`
     padding-inline: 0;
   `}
 `
-
-export const HeaderContainer = styled('header')`
-  display: flex;
-  align-items: center;
-  margin-bottom: 1em;
-  font-family: var(--font-feature);
-  gap: 1em;
-
-  ${p => p.$center && `
-    justify-content: center;
-  `}
-`
